+++ dds/DCPS/transport/framework/ThreadSynchStrategy.h
@@ -31,7 +31,7 @@ public:
   virtual ThreadSynch* create_synch_object(
     ThreadSynchResource* synch_resource,
     long                 priority,
-    int                  scheduler) = 0;
+    long                 scheduler) = 0;
 
 protected:
